refactor: abstract grid into Layout component
