+++ packages/website/pages/_app.tsx
@@ -1,26 +1,9 @@
 import React from 'react';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
-import { css } from 'emotion';
-import { GlobalStyles, Grid } from '@contentful/f36-components';
+import { GlobalStyles } from '@contentful/f36-components';
 
-const styles = {
-  grid: css({
-    height: '100vh',
-    overflow: 'hidden',
-    gridTemplateAreas: `"topbar topbar"
-    "sidemenu content"`,
-  }),
-  gridItem: css({
-    display: 'flex',
-    flexDirection: 'column',
-    overflow: 'auto',
-  }),
-  footer: css({
-    position: 'sticky',
-    top: '100vh',
-  }),
-};
+import { Layout } from '../components/Layout';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
@@ -35,24 +18,9 @@ function MyApp({ Component, pageProps }: AppProps) {
         <link rel="icon" href="/favicon.png" />
       </Head>
 
-      <Grid
-        className={styles.grid}
-        columns="320px auto"
-        rows="auto 1fr"
-        columnGap="none"
-      >
-        <Grid.Item area="topbar">[TOPBAR]</Grid.Item>
-
-        <Grid.Item area="sidemenu" className={styles.gridItem}>
-          [SideMenu]
-        </Grid.Item>
-
-        <Grid.Item area="content" as="main" className={styles.gridItem}>
-          <Component {...pageProps} />
-
-          <footer className={styles.footer}>[FOOTER]</footer>
-        </Grid.Item>
-      </Grid>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
     </>
   );
 }
